telephony-service/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
aiomysql
httpx
websockets
